+++ node/src/GlideClusterClient.ts
@@ -23,6 +23,7 @@ import {
     createEcho,
     createFlushAll,
     createFlushDB,
+    createFunctionFlush,
     createFunctionLoad,
     createInfo,
     createLolwut,
@@ -685,6 +686,31 @@ export class GlideClusterClient extends BaseClient {
     ): Promise<string> {
         return this.createWritePromise(
             createFunctionLoad(libraryCode, replace),
+            toProtobufRoute(route),
+        );
+    }
+
+    /**
+     * Deletes all function libraries.
+     *
+     * See https://valkey.io/commands/function-flush/ for details.
+     *
+     * since Valkey version 7.0.0.
+     *
+     * @param mode - The flushing mode, could be either {@link FlushMode.SYNC} or {@link FlushMode.ASYNC}.
+     * @param route - The command will be routed to all primary node, unless `route` is provided, in which
+     *   case the client will route the command to the nodes defined by `route`.
+     * @returns A simple OK response.
+     *
+     * @example
+     * ```typescript
+     * const result = await client.functionFlush(FlushMode.SYNC);
+     * console.log(result); // Output: 'OK'
+     * ```
+     */
+    public functionFlush(mode?: FlushMode, route?: Routes): Promise<string> {
+        return this.createWritePromise(
+            createFunctionFlush(mode),
             toProtobufRoute(route),
         );
     }